+++ cpp/include/cudf/ast/detail/expression_parser.hpp
@@ -30,7 +30,7 @@ namespace ast {
 namespace detail {
 
 /**
- * @brief Enum defining data reference types used by a node.
+ * @brief Node data reference types.
  *
  * This enum is device-specific. For instance, intermediate data references are generated by the
  * linearization process but cannot be explicitly created by the user.
