+++ cpp/tests/ast/transform_tests.cpp
@@ -15,7 +15,6 @@
  */
 
 #include <cudf/ast/operators.hpp>
-#include <cudf/ast/transform.hpp>
 #include <cudf/column/column.hpp>
 #include <cudf/column/column_view.hpp>
 #include <cudf/detail/iterator.cuh>
@@ -24,6 +23,7 @@
 #include <cudf/scalar/scalar_factories.hpp>
 #include <cudf/table/table.hpp>
 #include <cudf/table/table_view.hpp>
+#include <cudf/transform.hpp>
 #include <cudf/types.hpp>
 
 #include <cudf_test/base_fixture.hpp>
@@ -58,7 +58,7 @@ TEST_F(TransformTest, BasicAddition)
   auto expression = cudf::ast::expression(cudf::ast::ast_operator::ADD, col_ref_0, col_ref_1);
 
   auto expected = column_wrapper<int32_t>{13, 27, 21, 50};
-  auto result   = cudf::ast::compute_column(table, expression);
+  auto result   = cudf::compute_column(table, expression);
 
   cudf::test::expect_columns_equal(expected, result->view(), verbosity);
 }
@@ -74,7 +74,7 @@ TEST_F(TransformTest, BasicAdditionLarge)
 
   auto b        = cudf::detail::make_counting_transform_iterator(0, [](auto i) { return i * 2; });
   auto expected = column_wrapper<int32_t>(b, b + 2000);
-  auto result   = cudf::ast::compute_column(table, expression);
+  auto result   = cudf::compute_column(table, expression);
 
   cudf::test::expect_columns_equal(expected, result->view(), verbosity);
 }
@@ -90,7 +90,7 @@ TEST_F(TransformTest, LessComparator)
   auto expression = cudf::ast::expression(cudf::ast::ast_operator::LESS, col_ref_0, col_ref_1);
 
   auto expected = column_wrapper<bool>{true, false, true, false};
-  auto result   = cudf::ast::compute_column(table, expression);
+  auto result   = cudf::compute_column(table, expression);
 
   cudf::test::expect_columns_equal(expected, result->view(), verbosity);
 }
@@ -109,7 +109,7 @@ TEST_F(TransformTest, LessComparatorLarge)
 
   auto c        = cudf::detail::make_counting_transform_iterator(0, [](auto i) { return i < 500; });
   auto expected = column_wrapper<bool>(c, c + 2000);
-  auto result   = cudf::ast::compute_column(table, expression);
+  auto result   = cudf::compute_column(table, expression);
 
   cudf::test::expect_columns_equal(expected, result->view(), verbosity);
 }
@@ -134,7 +134,7 @@ TEST_F(TransformTest, MultiLevelTreeArithmetic)
   auto expression_tree = cudf::ast::expression(
     cudf::ast::ast_operator::ADD, expression_left_subtree, expression_right_subtree);
 
-  auto result   = cudf::ast::compute_column(table, expression_tree);
+  auto result   = cudf::compute_column(table, expression_tree);
   auto expected = column_wrapper<int32_t>{7, 73, 22, -99};
 
   cudf::test::expect_columns_equal(expected, result->view(), verbosity);
@@ -160,7 +160,7 @@ TEST_F(TransformTest, MultiLevelTreeArithmeticLarge)
   auto expr_right_subtree = expression(cudf::ast::ast_operator::ADD, col_ref_2, col_ref_0);
   auto expr_tree          = expression(ast_operator::SUB, expr_left_subtree, expr_right_subtree);
 
-  auto result = cudf::ast::compute_column(table, expr_tree);
+  auto result = cudf::compute_column(table, expr_tree);
   auto calc   = [](auto i) { return (i * (i + 1)) - (i + (i * 2)); };
   auto d      = cudf::detail::make_counting_transform_iterator(0, [&](auto i) { return calc(i); });
   auto expected = column_wrapper<int32_t>(d, d + 2000);
@@ -185,7 +185,7 @@ TEST_F(TransformTest, ImbalancedTreeArithmetic)
   auto expression_tree =
     cudf::ast::expression(cudf::ast::ast_operator::SUB, col_ref_2, expression_right_subtree);
 
-  auto result = cudf::ast::compute_column(table, expression_tree);
+  auto result = cudf::compute_column(table, expression_tree);
   auto expected =
     column_wrapper<double>{0.6, std::numeric_limits<double>::infinity(), -3.201, -2099.18};
 
@@ -212,7 +212,7 @@ TEST_F(TransformTest, MultiLevelTreeComparator)
   auto expression_tree = cudf::ast::expression(
     cudf::ast::ast_operator::LOGICAL_AND, expression_left_subtree, expression_right_subtree);
 
-  auto result   = cudf::ast::compute_column(table, expression_tree);
+  auto result   = cudf::compute_column(table, expression_tree);
   auto expected = column_wrapper<bool>{false, true, false, false};
 
   cudf::test::expect_columns_equal(expected, result->view(), verbosity);
@@ -233,8 +233,8 @@ TEST_F(TransformTest, MultiTypeOperationFailure)
     cudf::ast::expression(cudf::ast::ast_operator::ADD, col_ref_1, col_ref_0);
 
   // Operations on different types are not allowed
-  EXPECT_THROW(cudf::ast::compute_column(table, expression_0_plus_1), cudf::logic_error);
-  EXPECT_THROW(cudf::ast::compute_column(table, expression_1_plus_0), cudf::logic_error);
+  EXPECT_THROW(cudf::compute_column(table, expression_0_plus_1), cudf::logic_error);
+  EXPECT_THROW(cudf::compute_column(table, expression_1_plus_0), cudf::logic_error);
 }
 
 TEST_F(TransformTest, LiteralComparison)
@@ -248,7 +248,7 @@ TEST_F(TransformTest, LiteralComparison)
 
   auto expression = cudf::ast::expression(cudf::ast::ast_operator::GREATER, col_ref_0, literal);
 
-  auto result   = cudf::ast::compute_column(table, expression);
+  auto result   = cudf::compute_column(table, expression);
   auto expected = column_wrapper<bool>{false, false, false, true};
 
   cudf::test::expect_columns_equal(expected, result->view(), verbosity);
@@ -263,7 +263,7 @@ TEST_F(TransformTest, UnaryNot)
 
   auto expression = cudf::ast::expression(cudf::ast::ast_operator::NOT, col_ref_0);
 
-  auto result   = cudf::ast::compute_column(table, expression);
+  auto result   = cudf::compute_column(table, expression);
   auto expected = column_wrapper<bool>{false, true, false, false};
 
   cudf::test::expect_columns_equal(expected, result->view(), verbosity);
@@ -278,17 +278,17 @@ TEST_F(TransformTest, UnaryTrigonometry)
 
   auto expected_sin   = column_wrapper<double>{0.0, std::sqrt(2) / 2, std::sqrt(3.0) / 2.0};
   auto expression_sin = cudf::ast::expression(cudf::ast::ast_operator::SIN, col_ref_0);
-  auto result_sin     = cudf::ast::compute_column(table, expression_sin);
+  auto result_sin     = cudf::compute_column(table, expression_sin);
   cudf::test::expect_columns_equivalent(expected_sin, result_sin->view(), verbosity);
 
   auto expected_cos   = column_wrapper<double>{1.0, std::sqrt(2) / 2, 0.5};
   auto expression_cos = cudf::ast::expression(cudf::ast::ast_operator::COS, col_ref_0);
-  auto result_cos     = cudf::ast::compute_column(table, expression_cos);
+  auto result_cos     = cudf::compute_column(table, expression_cos);
   cudf::test::expect_columns_equivalent(expected_cos, result_cos->view(), verbosity);
 
   auto expected_tan   = column_wrapper<double>{0.0, 1.0, std::sqrt(3.0)};
   auto expression_tan = cudf::ast::expression(cudf::ast::ast_operator::TAN, col_ref_0);
-  auto result_tan     = cudf::ast::compute_column(table, expression_tan);
+  auto result_tan     = cudf::compute_column(table, expression_tan);
   cudf::test::expect_columns_equivalent(expected_tan, result_tan->view(), verbosity);
 }
 
@@ -311,7 +311,7 @@ TEST_F(TransformTest, StringComparison)
   auto expression = cudf::ast::expression(cudf::ast::ast_operator::LESS, col_ref_0, col_ref_1);
 
   auto expected = column_wrapper<bool>{true, false, true, false};
-  auto result   = cudf::ast::compute_column(table, expression);
+  auto result   = cudf::compute_column(table, expression);
 
   cudf::test::expect_columns_equal(expected, result->view(), verbosity);
 }
@@ -324,7 +324,7 @@ TEST_F(TransformTest, CopyColumn)
   auto col_ref_0  = cudf::ast::column_reference(0);
   auto expression = cudf::ast::expression(cudf::ast::ast_operator::IDENTITY, col_ref_0);
 
-  auto result   = cudf::ast::compute_column(table, expression);
+  auto result   = cudf::compute_column(table, expression);
   auto expected = column_wrapper<int32_t>{3, 0, 1, 50};
 
   cudf::test::expect_columns_equal(expected, result->view(), verbosity);
@@ -340,7 +340,7 @@ TEST_F(TransformTest, CopyLiteral)
 
   auto expression = cudf::ast::expression(cudf::ast::ast_operator::IDENTITY, literal);
 
-  auto result   = cudf::ast::compute_column(table, expression);
+  auto result   = cudf::compute_column(table, expression);
   auto expected = column_wrapper<int32_t>{-123, -123, -123, -123};
 
   cudf::test::expect_columns_equal(expected, result->view(), verbosity);
@@ -357,7 +357,7 @@ TEST_F(TransformTest, TrueDiv)
 
   auto expression = cudf::ast::expression(cudf::ast::ast_operator::TRUE_DIV, col_ref_0, literal);
 
-  auto result   = cudf::ast::compute_column(table, expression);
+  auto result   = cudf::compute_column(table, expression);
   auto expected = column_wrapper<double>{1.5, 0.0, 0.5, 25.0};
 
   cudf::test::expect_columns_equal(expected, result->view(), verbosity);
@@ -374,7 +374,7 @@ TEST_F(TransformTest, FloorDiv)
 
   auto expression = cudf::ast::expression(cudf::ast::ast_operator::FLOOR_DIV, col_ref_0, literal);
 
-  auto result   = cudf::ast::compute_column(table, expression);
+  auto result   = cudf::compute_column(table, expression);
   auto expected = column_wrapper<double>{1.0, 0.0, 0.0, 25.0};
 
   cudf::test::expect_columns_equal(expected, result->view(), verbosity);
@@ -391,7 +391,7 @@ TEST_F(TransformTest, Mod)
 
   auto expression = cudf::ast::expression(cudf::ast::ast_operator::MOD, col_ref_0, literal);
 
-  auto result   = cudf::ast::compute_column(table, expression);
+  auto result   = cudf::compute_column(table, expression);
   auto expected = column_wrapper<double>{1.0, 0.0, -1.0, 0.0};
 
   cudf::test::expect_columns_equal(expected, result->view(), verbosity);
@@ -408,7 +408,7 @@ TEST_F(TransformTest, PyMod)
 
   auto expression = cudf::ast::expression(cudf::ast::ast_operator::PYMOD, col_ref_0, literal);
 
-  auto result   = cudf::ast::compute_column(table, expression);
+  auto result   = cudf::compute_column(table, expression);
   auto expected = column_wrapper<double>{1.0, 0.0, 1.0, 0.0};
 
   cudf::test::expect_columns_equal(expected, result->view(), verbosity);
@@ -425,7 +425,7 @@ TEST_F(TransformTest, BasicAdditionNulls)
   auto expression = cudf::ast::expression(cudf::ast::ast_operator::ADD, col_ref_0, col_ref_1);
 
   auto expected = column_wrapper<int32_t>{{0, 0, 0, 50}, {0, 0, 0, 1}};
-  auto result   = cudf::ast::compute_column(table, expression);
+  auto result   = cudf::compute_column(table, expression);
 
   cudf::test::expect_columns_equal(expected, result->view(), verbosity);
 }
@@ -451,7 +451,7 @@ TEST_F(TransformTest, BasicAdditionLargeNulls)
 
   auto b        = cudf::detail::make_counting_transform_iterator(0, [](auto i) { return i * 2; });
   auto expected = column_wrapper<int32_t>(b, b + N, validities.begin());
-  auto result   = cudf::ast::compute_column(table, expression);
+  auto result   = cudf::compute_column(table, expression);
 
   cudf::test::expect_columns_equal(expected, result->view(), verbosity);
 }
